POC code development based on rev.1b - this commit introduces SD Card logging via FAT, a logger_t object with internal buffering
in autocharge.c/.h:
- refactored uint16_t _IsAdjustableLoadConnected(void) into a public uint16_t AdjustableLoadState
- refactored PowerMonitor to become a public member

in mj838_zerocross.c:
- adapted for above uint16_t AdjustableLoadState change

in mj838.c/.h:
- added logging of various data members
- added logging of AdjustableLoadState to existing logging

in sht40.c/.h:
- changed public measurement variables to be pointers to static (in .c file) variables

- converted relative humidity variable from int8_t to uint8_t

implemented mj838_SDCardParser.py: parses log outfile and imports it into an excel

in spi.c:
- switched to polling mode
- identified a bug/issue when SPI is powered off. a power cycle introduces delays into the whole system (visible in timer16cntr delay)
- tested IT-based transfer functions (seem to work)
- reduced polling timeout to 5
- corrected USE_SPI_POWER_MANAGEMENT across relevant parts of the code
- added USE_BUSCHECK, placed it around calls to _BusCheck() in the code

in diskio.c:
- in IO functions ignored BYTE pdrv, since we have only one disk

in ffconf.h:
- tuned off FF_USE_FASTSEEK
- deactivated FF_FS_LOCK

in user_diskio_spi.c:
- cleaned up USER_SPI_ioctl() and split it into helper functions
- cleaned up USER_SPI_write()
- cleaned up USER_SPI_read()
- cleaned up send_cmd()
- changed return type for rcvr_datablock() and xmit_datablock()
- changed return value type of SDCard_WaitForResponse() from uint8_t to SD_DataBusyReadyToken
- in rcvr_datablock() and xmit_datablock() added checks for above return value change
- added defines for SD Card Data Response Tokens
- added enums for status control: SD_DataResponseToken, SD_DataStartStopToken, SD_DataErrorToken, SD_DataBusyReadyToken
- used defined response tokens instead of raw hex numbers
- inlined SPI_Timer_On() and SPI_Timer_Status()
- in and around wait_ready(), despiselect() and spiselect(): simplified code a bit, added doxygen documentation
- renamed despiselect() to SDCardSPI_DeSelect()
- renamed spiselect() to SDCardSPI_Select()
- wrote SDCard_WaitForResponse() - a function which replaces the old wait_ready() and other parts of code where a SD card response & timeout was required
in user_diskio_spi.c:
- replaced HAL_SPI_X calls with Device->mj8x8->spi->X calls
- in function declarations declared arguments as const, where possible
- removed count sanity checks in USER_SPI_write() and USER_SPI_read()
- removed FCLK_SLOW() and FCLK_FAST: were originally used in USER_SPI_initialize()
- removed the BYTE pdrv argument from all USER_SPI_* functions
    in diskio.c:
        - implemented above chang
- generated doxygen documentation
- comment & format changes
- in rcvr_datablock() replaced call to "rcvr_spi_multi(buff, btr);" with for loop running "HAL_SPI_TransmitReceive(&_hspi1, &_dummy_spi_byte, buff + i, 1, 50);"
- in rcvr_spi_multi() replaced "*(buff + i) = xchg_spi(0xFF);" with "HAL_SPI_TransmitReceive(&_hspi1, &_dummy_spi_byte, buff + i, 1, 50);"
- in USER_SPI_initialize() replaced "ocr[n] = xchg_spi(0xFF);" with "HAL_SPI_TransmitReceive(&_hspi1, &_dummy_spi_byte, &ocr[n], 1, 50);"
- in"send_cmd() replaced  "res = xchg_spi(0xFF);" with "HAL_SPI_TransmitReceive(&_hspi1, &_dummy_spi_byte, &res, 1, 20);
- in rcvr_datablock() replaced "token = xchg_spi(0xFF)" with "HAL_SPI_TransmitReceive(&_hspi1, &_dummy_spi_byte, &token, 1, 20);"
- in wait_ready() replaced "d = xchg_spi(0xFF);" with "HAL_SPI_TransmitReceive(&_hspi1, &_dummy_spi_byte, &d, 1, 20);"
- in xmit_datablock() replaced "resp = xchg_spi(0xFF);" with "HAL_SPI_TransmitReceive(&_hspi1, &_dummy_spi_byte, &resp, 1, 20);"
- in xmit_datablock() replaced "xchg_spi(token);" with "HAL_SPI_TransmitReceive(&_hspi1, &token, &_rx, 1, 20);"
- in xmit_datablock() replaced Device->mj8x8->spi->Transmit() with HAL_SPI_Transmit()
- changed function declaration from "static BYTE send_cmd(BYTE cmd, DWORD arg)" to "static BYTE send_cmd(uint8_t cmd, uint32_t arg)"
- in send_cmd() changed txbyte data type from BYTE to uint8_t
- in send_cmd() replaced "Send command packet" occurances of xchg_spi() with HAL_SPI_TransmitReceive()
- replaced occurances of xchg_spi(0xFF); with HAL_SPI_TransmitReceive(&_hspi1, &_dummy_spi_byte, &_rx, 1, 20);
- converted despiselect()/spiselect() to HAL SPI
- replaced xmit_spi_multi() with Device->mj8x8->spi->Transmit((uint8_t*) buff, 512);

in sdcard.c/.h:
wrote SD card object & tied in https://elm-chan.org/fsw/ff/ FAT filesystem driver

in logger.c:
wrote logger object which utilizes ring buffer from https://docs.majerle.eu/projects/lwrb/en/latest/ and which utilizes above SD card

+++ mj8x8/Core/Inc/mj838/ina219/ina219.c
@@ -26,9 +26,9 @@ typedef struct	// ina219_t actual
 #if CALCULATE_SHUNT_VOLTAGE
 int16_t _ShuntVoltage;
 #endif
-uint16_t _Voltage;
-int16_t _Current;
-uint16_t _Power;
+uint32_t _Voltage;  // private variable for voltage
+int32_t _Current;  // private variable for current
+uint32_t _Power;	// private variable for power
 
 // DS. p. 16, figure 15 - write word command
 static inline void _WriteWord(const uint8_t RegAddr, uint8_t *data)
@@ -115,17 +115,16 @@ static inline void _Measure(void)
 
 	_RegisterPointerSet(BUS_VOLTAGE_REG);  // set pointer to voltage register
 	_ReadWord(temp);  // read 2 bytes from voltage register
-	_Voltage = (uint16_t) (((temp[0] << 8) | (temp[1])) >> 3) * 4;  // combine the two bytes into a single 16-bit value, shift away 3 LSBs
+	_Voltage = (uint32_t) ((((temp[0] << 8) | (temp[1])) >> 3) * 4);  // combine the two bytes into a single 16-bit value, shift away 3 LSBs
 
 	_RegisterPointerSet(CURRENT_REG);  // set pointer to current register
 	_ReadWord(temp);  // read 2 bytes from current register
-	_Current = (int16_t) ((_2sComplementTo_uint16(temp) * 61) / 1000);  // combine, scale, and convert to mA in one statement
+	_Current = (int32_t) ((_2sComplementTo_uint16(temp) * 61) / 1000);  // combine, scale, and convert to mA in one statement
 
 	_RegisterPointerSet(POWER_REG);  // set pointer to power register
 	_ReadWord(temp);  // read 2 bytes from power register
-//	_Power = (uint16_t) (((temp[0] << 8) | (temp[1])) * 1.220703125f);  // power LSB is 20 times the current LSB
-	// FIXME - validate correct values
-	_Power = (uint16_t) ((((temp[0] << 8) | (temp[1])) * 1250) >> 10);  // 1250/1024 ≈ 1.220703125f
+	// power LSB is 20 times the current LSB
+	_Power = (uint32_t) ((((temp[0] << 8) | (temp[1])) * 1250) >> 10);  // 1250/1024 ≈ 1.220703125f
 }
 
 static __ina219_t __INA219 __attribute__ ((section (".data"))) =  // instantiate ina219_t actual and set function pointers
